% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anva.R
\name{AnVa}
\alias{AnVa}
\title{Análisis de varianza}
\usage{
AnVa(w, x, y, z)
}
\arguments{
\item{w}{(tratamiento) datos del efecto del T1 sobre la(s) variable(s)}

\item{x}{(tratamiento) datos del efecto del T2 sobre la(s) variable(s)}

\item{y}{(tratamiento) datos del efecto del T3 sobre la(s) variable(s)}

\item{z}{(tratamiento) datos del efecto del T4 sobre la(s) variable(s)}
}
\value{
Una lista con la Media de los tratamientos, los Grados de libertad, la Suma de Cuadrados, el valor de F calculada y los Grados de libertad para el cálculo de F tablas
}
\description{
Realiza un análisis de varianza para datos completamente al azar con igual número de unidades por tratamiento
}
